Possibly fix deterministic (seeded) generation
On a Windows machine this code is sufficient for having reproducible results pixel-to-pixel.
Also tried the effect of np.random.seed() and random.seed() but they don't seem to have any
effect, at least on my configuration (PyTorch 1.7.1, CUDA 10.1, Windows x64).

This introduces a slowdown of ~6%, but honors the seed. Interpolation is not supported with
determinism because of the 'upsample_bilinear2d_backward_cuda' operation.

The code replaced (below) also seems to have some virtues as the style is generally very
similar, but there is some variability on the appearance:
  torch.cuda.manual_seed(seed)
  torch.backends.cudnn.deterministic=True
(the lines above are now replaced by torch.set_deterministic(True))